chore: fix esbuild

+++ esbuild.config.mjs
@@ -1,7 +1,11 @@
 import builtins from "builtin-modules";
+import dotenv from "dotenv";
 import esbuild from "esbuild";
 import * as fs from "fs";
+import * as path from "path";
 import process from "process";
+
+dotenv.config();
 
 const banner =
 `/*
@@ -16,7 +20,7 @@ const moveStyles = {
 	name: "move-styles",
 	setup(build) {
 		build.onEnd(() => {
-			fs.copyFileSync("plugin/styles.css", "./styles.css");
+			if (fs.existsSync("plugin/styles.css")) fs.copyFileSync("plugin/styles.css", "./styles.css");
 		});
 	}
 };
@@ -25,18 +29,23 @@ const exportToVaultFunc = {
 	name: "export-to-vault",
 	setup(build) {
 		build.onEnd(() => {
-			if (prod && exportToVault) {
-				const vaultPath = process.env.VAULT_PATH;
-				const pluginManifest = JSON.parse(fs.readFileSync("./manifest.json", "utf-8"));
-				const pluginId = pluginManifest.id;
-				const pluginFolder = `${vaultPath}/.obsidian/plugins/${pluginId}`;
-				if (!fs.existsSync(pluginFolder)) {
-					fs.mkdirSync(pluginFolder, {recursive: true});
-				}
-				fs.copyFileSync("./dist/main.js", `${pluginFolder}/main.js`);
-				fs.copyFileSync("./dist/styles.css", `${pluginFolder}/styles.css`);
-				fs.copyFileSync("./manifest.json", `${pluginFolder}/manifest.json`);
+			if (!(prod && exportToVault)) {
+				return;
 			}
+			if (!process.env.VAULT_PATH) {
+				console.error("VAULT_PATH environment variable not set, skipping export to vault");
+				return;
+			}
+			const vaultPath = process.env.VAULT_PATH;
+			const pluginManifest = JSON.parse(fs.readFileSync("./manifest.json", "utf-8"));
+			const pluginId = pluginManifest.id;
+			const pluginFolder = path.join(vaultPath,".obsidian", "plugins", pluginId);
+			if (!fs.existsSync(pluginFolder)) {
+				fs.mkdirSync(pluginFolder, {recursive: true});
+			}
+			fs.copyFileSync("./main.js", path.join(pluginFolder, "main.js"));
+			if (fs.existsSync("./styles.css")) fs.copyFileSync("./styles.css", path.join(pluginFolder, "styles.css"));
+			fs.copyFileSync("./manifest.json", path.join(pluginFolder, "manifest.json"));
 		});
 	}
 };
@@ -45,11 +54,16 @@ const exportToDist = {
 	name: "export-to-dist",
 	setup(build) {
 		build.onEnd(() => {
-			if (prod) {
-				fs.copyFileSync("./main.js", "./dist/main.js");
-				fs.copyFileSync("./styles.css", "./dist/styles.css");
-				fs.copyFileSync("./manifest.json", "./dist/manifest.json");
+			if (!prod) {
+				return;
 			}
+			if (!fs.existsSync("./dist")) {
+				fs.mkdirSync("./dist");
+			}
+			fs.copyFileSync("main.js", path.join("./dist", "main.js"));
+			if (fs.existsSync("styles.css"))
+				fs.copyFileSync("styles.css", path.join("./dist", "styles.css"));
+			fs.copyFileSync("manifest.json", path.join("./dist", "manifest.json"));
 		});
 	}
 };
